Bug fix
fixed a bug where times would reset to last saved time when the add task button was pressed.
It now saves the times for all the tasks to the task.json, before destroying the old task frame.
when the new task frame is generated it will pull the updated time.

Currently timer are all stoped when a new task is added